Added jest, moved types from signup to correctly build
