+++ package.json
@@ -7,6 +7,7 @@
 		"build": "next build",
 		"start": "next start",
 		"lint": "next lint",
+		"test": "jest",
 		"lint:fix": "next lint --fix",
 		"prisma:migrate": "yarn prisma migrate dev",
 		"prisma:migrate:deploy": "yarn prisma migrate deploy",
@@ -34,6 +35,8 @@
 	},
 	"devDependencies": {
 		"@next/eslint-plugin-next": "^12.2.5",
+		"@testing-library/jest-dom": "^5.16.5",
+		"@testing-library/react": "^13.3.0",
 		"@types/node": "^18.7.3",
 		"@types/react": "18.0.17",
 		"@typescript-eslint/eslint-plugin": "^5.33.0",
@@ -44,6 +47,8 @@
 		"eslint-plugin-react": "^7.30.1",
 		"eslint-plugin-react-hooks": "^4.6.0",
 		"husky": "^8.0.1",
+		"jest": "^28.1.3",
+		"jest-environment-jsdom": "^28.1.3",
 		"lint-staged": "^13.0.3",
 		"prettier": "^2.7.1",
 		"prisma": "^4.2.1",
