+++ pages/index.tsx
@@ -1,7 +1,7 @@
 // import { InitialState } from '@stores/types';
 // { user, session: userSession }: InitialState
 const Homepage = () => {
-	return <div>Root page</div>;
+	return <div data-testid='homepage-main'>Root page</div>;
 };
 
 export default Homepage;
