--- pages/signup/types.ts
@@ -1,5 +1,0 @@
-export interface RegisterData {
-	username: string;
-	password: string;
-	email: string;
-}
